django/requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
djangorestframework==3.9.3
fix-yahoo-finance==0.1.33
idna==2.7
kombu==4.5.0
lxml==4.3.3
multitasking==0.0.8
numpy==1.16.3
pandas==0.24.2
pandas-datareader==0.7.0
psycopg2==2.8.2
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.19.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.23
uWSGI==2.0.18
vine==1.3.0
wrapt==1.11.1
